+++ CHANGELOG.md
@@ -26,6 +26,11 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the
 
 ---
 
+### Unreleased - Added
+
+- [#211](https://github.com/ARGOeu/argo-accounting/pull/211) - ACC-319 Create 
+  an admin-only endpoint to update user information.
+
 ## 1.6.0 - 2025-03-13
 
 ---